+++ src/lib/hw/common/common.h
@@ -17,6 +17,6 @@ typedef struct common_hw_state_struct common_hw_state_struct_t;
 int common_ti_calc_sync_delay(uint32_t clkpos, unsigned* calced_delay);
 int common_print_registers_a8d16(uint32_t* regs, unsigned count, int loglevel);
 int common_spi_post(void* o, uint32_t* regs, unsigned count);
-int common_spi_get(void* o, uint16_t addr, uint16_t* out);
+int common_spi_get(void* o, uint32_t addr, uint16_t* out);
 
 #endif // HW_COMMON_H
